+++ rhine/src/FRP/Rhine/ClSF/Except.hs
@@ -5,7 +5,7 @@
 {- | This module provides exception handling, and thus control flow,
 to synchronous signal functions.
 
-The API presented here closely follows dunai's 'Control.Monad.Trans.MSF.Except',
+The API presented here closely follows @automaton@'s 'Data.Automaton.Trans.Except',
 and reexports everything needed from there.
 -}
 module FRP.Rhine.ClSF.Except (
@@ -14,7 +14,7 @@ module FRP.Rhine.ClSF.Except (
   safe,
   safely,
   exceptS,
-  runMSFExcept,
+  runAutomatonExcept,
   currentInput,
 )
 where
@@ -27,12 +27,9 @@ import Control.Monad.Trans.Class (lift)
 import Control.Monad.Trans.Except as X
 import Control.Monad.Trans.Reader
 
--- dunai
-import Control.Monad.Trans.MSF.Except hiding (once, once_, throwOn, throwOn', throwS, try)
-import Data.MonadicStreamFunction
-
--- TODO Find out whether there is a cleverer way to handle exports
-import Control.Monad.Trans.MSF.Except qualified as MSFE
+-- automaton
+import Data.Automaton.Trans.Except hiding (once, once_, throwOn, throwOn', throwS, try)
+import Data.Automaton.Trans.Except qualified as AutomatonE
 
 -- rhine
 import FRP.Rhine.ClSF.Core
@@ -46,11 +43,11 @@ throwS :: (Monad m) => ClSF (ExceptT e m) cl e a
 throwS = arrMCl throwE
 
 -- | Immediately throw the given exception.
-throw :: (Monad m) => e -> MSF (ExceptT e m) a b
+throw :: (Monad m) => e -> Automaton (ExceptT e m) a b
 throw = constM . throwE
 
 -- | Do not throw an exception.
-pass :: (Monad m) => MSF (ExceptT e m) a a
+pass :: (Monad m) => Automaton (ExceptT e m) a a
 pass = Category.id
 
 -- | Throw the given exception when the 'Bool' turns true.
@@ -90,19 +87,20 @@ throwMaybe = proc me -> case me of
 -- * Monad interface
 
 {- | A synchronous exception-throwing signal function.
-It is based on a @newtype@ from Dunai, 'MSFExcept',
+
+It is based on a @newtype@ from @automaton@, 'AutomatonExcept',
 to exhibit a monad interface /in the exception type/.
 `return` then corresponds to throwing an exception,
 and `(>>=)` is exception handling.
-(For more information, see the documentation of 'MSFExcept'.)
+(For more information, see the documentation of 'AutomatonExcept'.)
 
-* @m@:  The monad that the signal function may take side effects in
 * @cl@: The clock on which the signal function ticks
 * @a@:  The input type
 * @b@:  The output type
+* @m@:  The monad that the signal function may take side effects in
 * @e@:  The type of exceptions that can be thrown
 -}
-type ClSFExcept m cl a b e = MSFExcept (ReaderT (TimeInfo cl) m) a b e
+type ClSFExcept m cl a b e = AutomatonExcept (ReaderT (TimeInfo cl) m) a b e
 
 {- | A clock polymorphic 'ClSFExcept',
 or equivalently an exception-throwing behaviour.
@@ -116,20 +114,20 @@ type BehaviorFExcept m time a b e = BehaviourFExcept m time a b e
 
 -- | Leave the monad context, to use the 'ClSFExcept' as an 'Arrow'.
 runClSFExcept :: (Monad m) => ClSFExcept m cl a b e -> ClSF (ExceptT e m) cl a b
-runClSFExcept = morphS commuteExceptReader . runMSFExcept
+runClSFExcept = hoistS commuteExceptReader . runAutomatonExcept
 
 {- | Enter the monad context in the exception
    for 'ClSF's in the 'ExceptT' monad.
    The 'ClSF' will be run until it encounters an exception.
 -}
 try :: (Monad m) => ClSF (ExceptT e m) cl a b -> ClSFExcept m cl a b e
-try = MSFE.try . morphS commuteReaderExcept
+try = AutomatonE.try . hoistS commuteReaderExcept
 
 {- | Within the same tick, perform a monadic action,
    and immediately throw the value as an exception.
 -}
 once :: (Monad m) => (a -> m e) -> ClSFExcept m cl a b e
-once f = MSFE.once $ lift . f
+once f = AutomatonE.once $ lift . f
 
 -- | A variant of 'once' without input.
 once_ :: (Monad m) => m e -> ClSFExcept m cl a b e
@@ -139,4 +137,4 @@ once_ = once . const
    and then throws an exception.
 -}
 step :: (Monad m) => (a -> m (b, e)) -> ClSFExcept m cl a b e
-step f = MSFE.step $ lift . f
+step f = AutomatonE.step $ lift . f